Added default inner to cmaes
